Return an integer from `bp_get_new_group_id()`.
Sanitization techniques introduced in [7790] inadvertently caused
the value returned from this function to be a string, inconsistent
with the documentation and with previous behavior.

Fixes #7388.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11299 cdf35c40-ae34-48e0-9cc9-0c9da1808c22

+++ src/bp-groups/bp-groups-template.php
@@ -4799,7 +4799,7 @@ function bp_new_group_id() {
 		 *
 		 * @param int $new_group_id ID of the new group.
 		 */
-		return apply_filters( 'bp_get_new_group_id', $new_group_id );
+		return (int) apply_filters( 'bp_get_new_group_id', $new_group_id );
 	}
 
 /**
